requirements.txt
torch
torchvision
pyyaml
albumentations
opencv-python
numpy
Pillow
tqdm
